remove unnecessary checks
Remove `command -v program` checks if program for the alias is
installed. I already installed the program that's why I added the alias
for the program. I was just needlessly being careful I suppose.

Added comments to sort of categorize aliases
Added alias for fzf with fancy icons for prompt