perf: use `custom/ipc` instead of `custom/script`
The polybar modules `brightness` and `volume` are previously
`custom/script` modules that run periodically. The interval I set are
very short (`interval = 1`) so that these modules are updated
immediately when I adjust the volume and brightness. This is very
inefficient and bugs me but it worked for an initial configuration of
polybar and did not significantly affect the performance of my system.
Also, I did not understand how to set up `custom/ipc` when I first
read about it XD.

It turns out `custom/ipc` is very easy to setup and I just have to use
 the `polybar-msg` command that I added to my keybindings in `sxhkdrc`