refactor: use a shell script for the time module
`exec` for the time module is set to `date +%H:%M` and that's really all
the module needs. However, I am using the builtin battery module and
upon reading the documentation, it looks like it doesn't have a way to
send a notification (when the battery level is low or full). So I'm
going to use the time module to run a script that outputs the time but
also checks for the battery level and sends notifications. I know this
is not the best way to do it but it will serve me well for now.