perf: only define `__prompt_command` if not in linux console
This probably does not matter in the grand scheme of things however it
seems logical to only define the prompt command after it is confirmed
the session is not in the linux console. I could even get rid of the
check but I do like to login only to the linux console sometimes and I
don't like to load the custom prompt in that session.