add rules for other scratchpad classes
I don't know why but when I'm launching a normal scratchpad the rule for
it to be a floating window is followed. But when I launch a scratchpad
with a running program the rule does not work anymore. As a (temporary?)
solution I'm just gonna add window rules for scratchpads that also
launch programs.